+++ kit/dapp/src/components/blocks/create-forms/bond/steps/basics.tsx
@@ -1,9 +1,9 @@
 "use client";
 
-import { StepContent } from "@/components/blocks/asset-designer/step-wizard/step-content";
 import type { AssetFormStep } from "@/components/blocks/asset-designer/types";
 import { FormStep } from "@/components/blocks/form/form-step";
 import { FormInput } from "@/components/blocks/form/inputs/form-input";
+import { StepContent } from "@/components/blocks/step-wizard/step-content";
 import type { CreateBondInput } from "@/lib/mutations/bond/create/create-schema";
 import { useTranslations } from "next-intl";
 import { useFormContext } from "react-hook-form";
