+++ kit/dapp/src/components/blocks/create-forms/common/asset-admins/asset-admins.tsx
@@ -1,10 +1,10 @@
 "use client";
 
-import { StepContent } from "@/components/blocks/asset-designer/step-wizard/step-content";
 import type { AssetFormStep } from "@/components/blocks/asset-designer/types";
 import { FormStep } from "@/components/blocks/form/form-step";
 import { FormInput } from "@/components/blocks/form/inputs/form-input";
 import { FormUsers } from "@/components/blocks/form/inputs/form-users";
+import { StepContent } from "@/components/blocks/step-wizard/step-content";
 import type { User } from "@/lib/queries/user/user-schema";
 import { useTranslations } from "next-intl";
 import { useEffect, useMemo, useState } from "react";
