+++ kit/dapp/src/components/blocks/create-forms/cryptocurrency/steps/configuration.tsx
@@ -1,9 +1,10 @@
 "use client";
 
-import { StepContent } from "@/components/blocks/asset-designer/step-wizard/step-content";
+import type { AssetFormStep } from "@/components/blocks/asset-designer/types";
 import { FormStep } from "@/components/blocks/form/form-step";
 import { FormInput } from "@/components/blocks/form/inputs/form-input";
 import { FormSelect } from "@/components/blocks/form/inputs/form-select";
+import { StepContent } from "@/components/blocks/step-wizard/step-content";
 import type { CreateCryptoCurrencyInput } from "@/lib/mutations/cryptocurrency/create/create-schema";
 import { fiatCurrencies } from "@/lib/utils/typebox/fiat-currency";
 import { useTranslations } from "next-intl";
@@ -88,7 +89,9 @@ Configuration.validatedFields = [
 ] satisfies (keyof CreateCryptoCurrencyInput)[];
 
 // Export step definition for the asset designer
-export const stepDefinition = {
+export const stepDefinition: AssetFormStep & {
+  component: typeof Configuration;
+} = {
   id: "configuration",
   title: "configuration.cryptocurrencies.title",
   description: "configuration.cryptocurrencies.description",
