+++ kit/dapp/src/components/blocks/create-forms/deposit/steps/basics.tsx
@@ -1,8 +1,9 @@
 "use client";
 
-import { StepContent } from "@/components/blocks/asset-designer/step-wizard/step-content";
+import type { AssetFormStep } from "@/components/blocks/asset-designer/types";
 import { FormStep } from "@/components/blocks/form/form-step";
 import { FormInput } from "@/components/blocks/form/inputs/form-input";
+import { StepContent } from "@/components/blocks/step-wizard/step-content";
 import type { CreateDepositInput } from "@/lib/mutations/deposit/create/create-schema";
 import { useTranslations } from "next-intl";
 import { useFormContext } from "react-hook-form";
@@ -89,7 +90,9 @@ Basics.validatedFields = [
 ] satisfies (keyof CreateDepositInput)[];
 
 // Export step definition for the asset designer
-export const stepDefinition = {
+export const stepDefinition: AssetFormStep & {
+  component: typeof Basics;
+} = {
   id: "details",
   title: "basics.title",
   description: "basics.description",
