+++ kit/dapp/src/components/blocks/create-forms/deposit/steps/configuration.tsx
@@ -1,9 +1,10 @@
 "use client";
 
-import { StepContent } from "@/components/blocks/asset-designer/step-wizard/step-content";
+import type { AssetFormStep } from "@/components/blocks/asset-designer/types";
 import { FormStep } from "@/components/blocks/form/form-step";
 import { FormInput } from "@/components/blocks/form/inputs/form-input";
 import { FormSelect } from "@/components/blocks/form/inputs/form-select";
+import { StepContent } from "@/components/blocks/step-wizard/step-content";
 import type { CreateDepositInput } from "@/lib/mutations/deposit/create/create-schema";
 import { fiatCurrencies } from "@/lib/utils/typebox/fiat-currency";
 import { timeUnits } from "@/lib/utils/typebox/time-units";
@@ -115,7 +116,9 @@ Configuration.validatedFields = [
 ] satisfies (keyof CreateDepositInput)[];
 
 // Export step definition for the asset designer
-export const stepDefinition = {
+export const stepDefinition: AssetFormStep & {
+  component: typeof Configuration;
+} = {
   id: "configuration",
   title: "configuration.deposits.title",
   description: "configuration.deposits.description",
