+++ kit/dapp/src/components/blocks/create-forms/stablecoin/steps/configuration.tsx
@@ -1,7 +1,8 @@
-import { StepContent } from "@/components/blocks/asset-designer/step-wizard/step-content";
+import type { AssetFormStep } from "@/components/blocks/asset-designer/types";
 import { FormStep } from "@/components/blocks/form/form-step";
 import { FormInput } from "@/components/blocks/form/inputs/form-input";
 import { FormSelect } from "@/components/blocks/form/inputs/form-select";
+import { StepContent } from "@/components/blocks/step-wizard/step-content";
 import type { CreateStablecoinInput } from "@/lib/mutations/stablecoin/create/create-schema";
 import { fiatCurrencies } from "@/lib/utils/typebox/fiat-currency";
 import { timeUnits } from "@/lib/utils/typebox/time-units";
@@ -113,7 +114,9 @@ Configuration.validatedFields = [
 ] satisfies (keyof CreateStablecoinInput)[];
 
 // Export step definition for the asset designer
-export const stepDefinition = {
+export const stepDefinition: AssetFormStep & {
+  component: typeof Configuration;
+} = {
   id: "configuration",
   title: "configuration.stablecoins.title",
   description: "configuration.stablecoins.description",
