+++ kit/dapp/src/components/blocks/step-wizard/mini-progress-bar.tsx
